Created "DiscountGrpcService" Class.
This class provides an abstraction on accessing the discount grpc client, so we can invoke the methods provided in this class from anywhere without knowing how communication between the grpc client and server is done.
In my personal opinion, the idea of this class is smart but it should've been implemented in this way, first we create an interface with the name "IDiscountGrpcClientService" which defines the methods that a grpc client would need, then we implement this interface by this class where the name can be set to match the interface name.
In order to keep consistency with the course, the implementation was left as it was introduced in the course.